MB-64957 do not record same slot-replica pair on multiple nodes
bg: it can happen that if a bad placement is used, we can end up
in a situation where some slot-replica end up on the multiple nodes;
this is illegal and can lead to multiple issues with rebalance/create
index.

exp: if there is an attempt to record the same slot-replica pair on
multiple nodes then fail the record request. if this request is made
during assignment of new shards then we should also be able to revert
the shard assigment for other replicas.

assumption: planner layout is such that same partition (same instance)
is going to multiple nodes.

Change-Id: Ib04e362fc79f6795bbf06396747ebd51bc18aceb
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>